+++ ds1961.h
@@ -3,6 +3,6 @@
 
 #include "OneWire.h"
 
-bool ReadAuthWithChallenge(OneWire *ow, const uint8_t id[], int addr, const uint8_t challenge[], uint8_t data[], uint8_t mac[]);
+bool ReadAuthWithChallenge(OneWire *ow, const uint8_t id[], uint16_t addr, const uint8_t challenge[], uint8_t data[], uint8_t mac[]);
 bool DS1961WriteSecret(OneWire *ow, const uint8_t id[], const uint8_t secret[]);
